Fjerner egen stegvelger.tsx i unntaksregistrering-flyten. Dette er stegvelgeren EnkelStegvelger baserer seg på så var ikke mye som måtte endres. Dro ut stegene i initialStegArray.tsx tilsvarende de andre flytene. vurderingInngang.tsx-steget henter oppfriskOgLastInnSaksopplysninger direkte fra fellesHandlersContext direkte istedenfor å gå gjennom saksbehandling.tsx og stegvelger.tsx
